+++ crates/asm/subprotocols/upgrade/src/actions/enact.rs
@@ -4,21 +4,21 @@ use crate::{
     error::UpgradeError,
     multisig::{msg::MultisigOp, vote::AggregatedVote},
     state::UpgradeSubprotoState,
-    txs::cancel::CancelAction,
+    txs::enact::EnactAction,
 };
 
 /// Handles a CancelAction transaction. It validates the vote on the cancellation
 /// and, if valid, removes the specified pending action from the state.
-pub fn handle_cancel_tx(
+pub fn handle_enactment_tx(
     state: &mut UpgradeSubprotoState,
     tx: &TxInput<'_>,
 ) -> Result<(), UpgradeError> {
     // Extract the aggregated vote and CancelAction from the transaction payload
     let vote = AggregatedVote::extract_from_tx(tx)?;
-    let cancel_action = CancelAction::extract_from_tx(tx)?;
+    let enact_action = EnactAction::extract_from_tx(tx)?;
 
     // Determine the ID of the pending action that should be canceled
-    let target_action_id = *cancel_action.id();
+    let target_action_id = *enact_action.id();
     let pending_action = state
         .get_pending_action(&target_action_id)
         .ok_or(UpgradeError::UnknownAction(target_action_id))?;
@@ -30,7 +30,7 @@ pub fn handle_cancel_tx(
         .ok_or(UpgradeError::UnknownRole)?;
 
     // Convert the cancel action into a multisig operation and validate it against the vote
-    let op = MultisigOp::from(cancel_action);
+    let op = MultisigOp::from(enact_action);
     authority.validate_op(&op, &vote)?;
 
     // All checks passed—remove the pending action from the state
