+++ crates/asm/subprotocols/upgrade/src/multisig/msg.rs
@@ -1,12 +1,14 @@
 use borsh::{BorshDeserialize, BorshSerialize};
 
-use crate::actions::{cancel::CancelAction, upgrades::UpgradeAction};
+use crate::txs::{cancel::CancelAction, enact::EnactAction, updates::UpgradeAction};
 
 /// A high‐level multisig operation that participants can propose.
 #[derive(Debug, Clone, Eq, PartialEq, BorshSerialize, BorshDeserialize)]
 pub enum MultisigOp {
     /// Cancel a pending action
     Cancel(CancelAction),
+    /// Cancel a committed action
+    Enact(EnactAction),
     /// Propose an upgrade.
     Upgrade(UpgradeAction),
 }
@@ -14,7 +16,7 @@ pub enum MultisigOp {
 /// A multisig payload comprising an operation plus a nonce, ready for hashing and signing.
 #[derive(Debug, Clone, Eq, PartialEq, BorshSerialize, BorshDeserialize)]
 pub struct MultisigPayload {
-    /// The multisig operation to execute (e.g., Cancel or Upgrade).
+    /// The multisig operation to execute.
     op: MultisigOp,
     /// A strictly increasing nonce to thwart replay.  
     nonce: u64,
@@ -35,5 +37,11 @@ impl From<UpgradeAction> for MultisigOp {
 impl From<CancelAction> for MultisigOp {
     fn from(cancel: CancelAction) -> Self {
         MultisigOp::Cancel(cancel)
+    }
+}
+
+impl From<EnactAction> for MultisigOp {
+    fn from(enact: EnactAction) -> Self {
+        MultisigOp::Enact(enact)
     }
 }
