+++ crates/asm/subprotocols/upgrade/src/multisig/vote.rs
@@ -1,11 +1,7 @@
 use borsh::{BorshDeserialize, BorshSerialize};
 use strata_asm_common::TxInput;
 
-use crate::{
-    actions::upgrades::UpgradeAction,
-    crypto::{PubKey, Signature, aggregate_pubkeys, verify_sig},
-    error::{DeserializeError, VoteValidationError},
-};
+use crate::{crypto::Signature, error::DeserializeError};
 
 /// An aggregated signature over a subset of signers in a MultisigConfig,
 /// identified by their positions in the config’s key list.
@@ -29,44 +25,6 @@ impl AggregatedVote {
 
     pub fn voter_indices(&self) -> &[u8] {
         &self.voter_indices
-    }
-
-    /// Validates this aggregated vote against the provided signer public keys for the given
-    /// `action_id`.
-    ///
-    /// This method performs three steps:
-    /// 1. Collect the individual public keys corresponding to `voter_indices` from `signers`.
-    /// 2. Aggregate those public keys into a single `PubKey` using `aggregate_pubkeys`.
-    /// 3. Verify the aggregated signature against the aggregated public key
-    pub fn validate_action(
-        &self,
-        signers: &[PubKey],
-        action: &UpgradeAction,
-    ) -> Result<(), VoteValidationError> {
-        // 1. Collect each public key by index; error if out of bounds.
-        let signer_keys: Vec<PubKey> = self
-            .voter_indices
-            .iter()
-            .map(|&i| {
-                signers
-                    .get(i as usize)
-                    .cloned()
-                    .ok_or(VoteValidationError::AggregationError)
-            })
-            .collect::<Result<_, _>>()?;
-
-        // 2. Aggregate those public keys into one.
-        let aggregated_key = aggregate_pubkeys(&signer_keys)?;
-
-        // 3. Compute the action ID from the UpgradeAction
-        let msg_hash = action.compute_id().into();
-
-        // 3. Verify the aggregated signature against the aggregated pubkey
-        if !verify_sig(&aggregated_key, &msg_hash, &self.agg_signature) {
-            return Err(VoteValidationError::InvalidVoteSignature);
-        }
-
-        Ok(())
     }
 }
